(HC-22) Port ConfigParseOptions
For now we leave config_includer unimplemented, as it's not being
exercised.

config_parse_options was kept an immutable object because we already
wanted to use a shared_ptr for config_includer, to avoid having to make
it a concrete class, and it keeps the implementation more in-line with
the original.

Copies should be light-weight, but we may reconsider and make this a
mutable class to avoid lots of copies if it's never really shared.