(HC-39) Port header files for SimpleConfigList
In order to port SimpleConfig List, add all the header files that
are needed. This includes the header file for SimpleConfigList and
several other classes that it relies on. At the moment, the header
files for resolve_context, resolve_result, and resolve_source are
just stubs without any methods or instance variables.