(HC-53) Finish enough of config_parser for simple cases
Finish enough of the config_parser that we can test simple hierarchies.

Fixed the config_object constructor, which was incorrectly trying to use
shared_from_this during construction. A config object is light-weight,
so create them as needed. If we need to compare two configs eventually,
we'll need to override them to do comparison of their roots instead.