(HC-52) Add basic config substitution tests
Add some basic tests for substitution and the relevant helper methods.
Also do a bit of code cleanup including making the needed operators
public and replacing some global instance variables with functions
instead (to avoid and error where the variable is not initialized
at the time the tests are run).