(HC-55) Implement a good chunk of the concatenation tests
This implements all of the string concatenation tests, and some of the
array and object tests. The full test suite requires us to have all
the accessors for lists and objects implemented, which we don't
yet. Since we're at the end of a sprint, getting this subset done
seemed expedient.