(FACT-1456) Add utility to wrap hocon in boost::program_options
This adds a `hocon::program_options` namespace with a `parse_hocon`
function. This function can be used the same as any of the normal
`program_options` parsers to generate config values, except it takes
as its input a hocon `shared_config` object.

HOCON allows representing certain data structures that
`boost::program_options` can not - specifically hashes inside arrays
cause trouble. We try to reprsent these when possible, (using a syntax
that treats the array index as a "key"), but defining actual
`po::options_description` sets that can represent these is Hard (if
not impossible).