(maint) System Locales fails regex backup
In case the locales on the system fails(due to gconv, or somethingelse),
leatherman will use boost::format and boost::regex, to create %N%style
formatting. 'regex' is only included in CMakeLists.txt if localesare
enabled, but if they fail, 'regex' will be required

This PR adds 'regex' to BOOST_COMPONENTS even if leatherman useslocales,
as an extra check in case the locales are enabled but fail.