Fix example and segfault when adding a value to an empty document
Fixes the example and makes it clearer how to compile it.

Also fixes the segfault when adding a value to an empty document. If
`children().empty()` is true, `children().front()` segfaults because
there's nothing for it to return a ref to. Rework logic so we only
attempt to check the first element if we know it exists.

Fixes #130.